+++ src/plone/restapi/tests/test_services_types.py
@@ -112,6 +112,11 @@ class TestServicesTypes(unittest.TestCase):
         self.assertIn(
             'file.data', response['properties']['file']['properties'])
 
+    def test_collection_type(self):
+        response = self.api_session.get('/@types/Collection')
+        response = response.json()
+        self.assertIn('anyOf', response['properties']['query']['items'])
+
     def test_addable_types_for_non_manager_user(self):
         user = api.user.create(
             email='noam.chomsky@example.com',